fix(rpc/v06): fix price unit in transaction receipt
`starknet_getTransactionReceipt` was still returning STRK as the fee
token type for v3 transactions.
